Create Fresh Climate
This script fit with Sentinel 2 L1. 
He always show the endothermic photosynthetic activity and reflective low clouds associated .

Albedo is calculated with RGB, Infrared B10 is necessary for temperature.
NDVI is on the green chanel and a little on blue to have a cold sensation.
Then Albedo is put on the blue channel too, because it is freshing when growing .
As Temperature is hot it is show on red channel .
So clouds are automaticaly analogic painted because they albedo and tempearture change.
It use square root function to well maximise the small temperature difference betwen environnement and barre soil, because high clouds seems to be very hot!!
As we maximise temperaure, note there is no red in the healty plant area...  

Will photosynthétic activity cooling the planet?
And note that blue clouds, because coldest with the highets albedos are always on the green healty plants area too.

Agroclimatics considarations:
Because when soil is brooken, there is no infiltation. 
So when there is less water,  vapor have to go upper to found fresh area in altitude an make a cloud. Or there is simply no clouds, so direct sunshine energy.
Bare soils are hots, beacause no endodthermic activity.
High clouds are hot because of little density making a tiny albedo, and this make easy to have much free water gases in atmosphere.(with is greenhouse gas activity)

So water have to condense as fast as possible, endothermic parts do that, and water have to stay in soil, not in the river! 
And as you know with much water plants do more time endothermic reaction, to capture rain, and puting enouth water to saturate, and condense in the first atmospheric layer.